Update: 파일 삭제 기능 구현 + 공지사항 수정 시 서버 파일 삭제 구현

+++ src/main/java/com/goodee/finals/notice/NoticeService.java
@@ -99,7 +99,11 @@ public class NoticeService {
 		
 		if (deleteFiles != null && deleteFiles.size() > 0) {
 			for (Long attachNum : deleteFiles) {
-				attachmentRepository.deleteById(attachNum);
+				String savedName = attachmentRepository.findById(attachNum).get().getSavedName();
+				boolean deleteResult = fileService.fileDelete(FileService.NOTICE, savedName);
+				if (deleteResult) {					
+					attachmentRepository.deleteById(attachNum);
+				}
 			}
 		}
